+++ install/upgrade/versions/1.4.4.sh
@@ -6,6 +6,12 @@
 #######                      Place additional commands below.                   #######
 #######################################################################################
 
+
+if [ -d "/etc/nginx/conf.d/" ]; then
+    #Add nginx user_agent separation to desktop/mobile
+    cp -f $HESTIA_INSTALL_DIR/nginx/agents.conf /etc/nginx/conf.d/
+fi
+
 # Reset PMA SSO to fix bug with Nginx + Apache2 
 if [ "$PHPMYADMIN_KEY" != "" ]; then
     echo "[ * ] Refressh hestia-sso for PMA..."